Join modifications from branch BR_DEBUG_3_2_0b1

+++ src/SALOMELocalTrace/Test/SALOMELocalTraceTest.cxx
@@ -15,7 +15,7 @@
 // License along with this library; if not, write to the Free Software 
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307 USA
 //
-// See http://www.salome-platform.org/
+// See http://www.salome-platform.org/ or email : webmaster.salome@opencascade.com
 //
 
 #include "SALOMELocalTraceTest.hxx"
@@ -180,9 +180,9 @@ SALOMELocalTraceTest::testLoadBufferPoolFile()
 
 void *PrintHello(void *threadid)
 {
-  int id_thread = (int)threadid;
+  long id_thread = (long)threadid;
   for (int i=0; i<NUM_MESSAGES;i++)
-    MESSAGE("Hello World! This is a trace test : " << id_thread 
-	    << " - iter " << i);
+  MESSAGE("Hello World! This is a trace test : " << id_thread 
+        << " - iter " << i);
   pthread_exit(NULL);
 }
